download data file from https://github.com/chenditc/investment_data

+++ qlib/tests/data.py
@@ -6,6 +6,7 @@ import re
 import sys
 import qlib
 import shutil
+import tarfile
 import zipfile
 import requests
 import datetime
@@ -16,7 +17,8 @@ from qlib.utils import exists_qlib_data
 
 
 class GetData:
-    REMOTE_URL = "https://github.com/SunsetWolf/qlib_dataset/releases/download"
+    REMOTE_DATA_URL = "https://github.com/SunsetWolf/qlib_dataset/releases/download"
+    REMOTE_BIN_URL = "https://github.com/chenditc/investment_data/releases"
 
     def __init__(self, delete_zip_file=False):
         """
@@ -28,7 +30,7 @@ class GetData:
         """
         self.delete_zip_file = delete_zip_file
 
-    def merge_remote_url(self, file_name: str):
+    def merge_remote_url(self, base_url: str, file_name: str):
         """
         Generate download links.
 
@@ -39,9 +41,9 @@ class GetData:
             The file name can be accompanied by a version number, (e.g.: v2/qlib_data_simple_cn_1d_latest.zip),
             if no version number is attached, it will be downloaded from v0 by default.
         """
-        return f"{self.REMOTE_URL}/{file_name}" if "/" in file_name else f"{self.REMOTE_URL}/v0/{file_name}"
+        return f"{base_url}/{file_name}" if "/" in file_name else f"{base_url}/v0/{file_name}"
 
-    def download(self, url: str, target_path: [Path, str]):
+    def download(self, url: str, target_path: Path, force_download: bool):
         """
         Download a file from the specified url.
 
@@ -52,24 +54,38 @@ class GetData:
         target_path: str
             The location where the data is saved, including the file name.
         """
-        file_name = str(target_path).rsplit("/", maxsplit=1)[-1]
-        resp = requests.get(url, stream=True, timeout=60)
-        resp.raise_for_status()
-        if resp.status_code != 200:
-            raise requests.exceptions.HTTPError()
+        if not target_path.exists() or force_download:
+            file_name = str(target_path).rsplit("/", maxsplit=1)[-1]
+            resp = requests.get(url, stream=True, timeout=60)
+            resp.raise_for_status()
+            if resp.status_code != 200:
+                raise requests.exceptions.HTTPError()
 
-        chunk_size = 1024
-        logger.warning(
-            f"The data for the example is collected from Yahoo Finance. Please be aware that the quality of the data might not be perfect. (You can refer to the original data source: https://finance.yahoo.com/lookup.)"
-        )
-        logger.info(f"{os.path.basename(file_name)} downloading......")
-        with tqdm(total=int(resp.headers.get("Content-Length", 0))) as p_bar:
-            with target_path.open("wb") as fp:
-                for chunk in resp.iter_content(chunk_size=chunk_size):
-                    fp.write(chunk)
-                    p_bar.update(chunk_size)
+            chunk_size = 1024
+            logger.warning(
+                f"The data for the example is collected from Yahoo Finance. Please be aware that the quality of the data might not be perfect. (You can refer to the original data source: https://finance.yahoo.com/lookup.)"
+            )
+            logger.info(f"{os.path.basename(file_name)} downloading......")
+            with tqdm(total=int(resp.headers.get("Content-Length", 0))) as p_bar:
+                with target_path.open("wb") as fp:
+                    for chunk in resp.iter_content(chunk_size=chunk_size):
+                        fp.write(chunk)
+                        p_bar.update(chunk_size)
 
-    def download_data(self, file_name: str, target_dir: [Path, str], delete_old: bool = True):
+    def _get_bin_file_name_with_version(self, name, bin_version, dataset_version):
+        if bin_version == "latest":
+            # "https://github.com/chenditc/investment_data/releases/latest/download/qlib_bin.tar.gz"
+            return  f"{bin_version}/download/{name}.tar.gz"
+        else:
+            # "https://github.com/chenditc/investment_data/releases/download/2020-10-11/qlib_bin.tar.gz"
+            return f"download/{bin_version}/{name}.tar.gz"
+
+    def _get_data_file_name_with_version(self, name, data_version, dataset_version):
+        dataset_version = "v2" if dataset_version is None else dataset_version
+        file_name_with_version = f"{dataset_version}/{name}_{self.region.lower()}_{self.interval.lower()}_{data_version}.zip"
+        return file_name_with_version
+
+    def download_data(self, name: str, target_dir: Path|str, delete_old: bool = True, force_download=True):
         """
         Download the specified file to the target folder.
 
@@ -82,6 +98,8 @@ class GetData:
             may contain folder names, for example: v2/qlib_data_simple_cn_1d_latest.zip
         delete_old: bool
             delete an existing directory, by default True
+        force_download: bool
+            download target file in force, by default True
 
         Examples
         ---------
@@ -97,27 +115,48 @@ class GetData:
         """
         target_dir = Path(target_dir).expanduser()
         target_dir.mkdir(exist_ok=True, parents=True)
-        # saved file name
-        _target_file_name = datetime.datetime.now().strftime("%Y%m%d%H%M%S") + "_" + os.path.basename(file_name)
-        target_path = target_dir.joinpath(_target_file_name)
+        get_file_name_with_version =  None
+        data_version: str = ''
+        today_str = datetime.datetime.now().strftime("%Y-%m-%d")
+        base_url: str = ''
 
-        url = self.merge_remote_url(file_name)
-        self.download(url=url, target_path=target_path)
+        if name == "qlib_bin":
+            data_version = today_str
+            get_file_name_with_version = self._get_bin_file_name_with_version
+            force_download = False
+            base_url = self.REMOTE_BIN_URL
+        else:
+            data_version = ".".join(re.findall(r"(\d+)\.+", qlib.__version__))
+            get_file_name_with_version = self._get_data_file_name_with_version
+            base_url = self.REMOTE_DATA_URL
+
+        url_path_file_name = get_file_name_with_version(name, data_version, dataset_version=self.version)
+        url = self.merge_remote_url(base_url, url_path_file_name)
+        target_file_name = f'{today_str}_{os.path.basename(url_path_file_name)}'
+        if not self.check_dataset(url):
+            force_download = True
+            url_path_file_name = get_file_name_with_version(name, "latest", dataset_version=self.version)
+            url = self.merge_remote_url(base_url, url_path_file_name)
+            target_file_name = f'latest_{os.path.basename(url_path_file_name)}'
+
+        # saved file name
+        target_path = target_dir.joinpath(target_file_name)
+
+        self.download(url=url, target_path=target_path, force_download=force_download)
 
         self._unzip(target_path, target_dir, delete_old)
         if self.delete_zip_file:
             target_path.unlink()
 
-    def check_dataset(self, file_name: str):
-        url = self.merge_remote_url(file_name)
-        resp = requests.get(url, stream=True, timeout=60)
+    def check_dataset(self, file_url: str):
+        resp = requests.get(file_url, stream=True, timeout=60)
         status = True
         if resp.status_code == 404:
             status = False
         return status
 
     @staticmethod
-    def _unzip(file_path: [Path, str], target_dir: [Path, str], delete_old: bool = True):
+    def _unzip(file_path: Path, target_dir: Path, delete_old: bool = True):
         file_path = Path(file_path)
         target_dir = Path(target_dir)
         if delete_old:
@@ -126,9 +165,17 @@ class GetData:
             )
             GetData._delete_qlib_data(target_dir)
         logger.info(f"{file_path} unzipping......")
-        with zipfile.ZipFile(str(file_path.resolve()), "r") as zp:
-            for _file in tqdm(zp.namelist()):
-                zp.extract(_file, str(target_dir.resolve()))
+        if file_path.suffix == ".gz":
+            # extract tar.gz file with strip-components=1 flag
+            with tarfile.open(str(file_path.resolve()), "r:gz") as tf:
+                members = tf.getmembers()
+                for member in tqdm(members):
+                    member.path = "/".join(member.path.split("/")[1:])  # strip-components=1
+                    tf.extract(member, str(target_dir.resolve()))
+        elif file_path.suffix == ".zip":
+            with zipfile.ZipFile(str(file_path.resolve()), "r") as zp:
+                for _file in tqdm(zp.namelist()):
+                    zp.extract(_file, str(target_dir.resolve()))
 
     @staticmethod
     def _delete_qlib_data(file_dir: Path):
@@ -159,6 +206,7 @@ class GetData:
         region="cn",
         delete_old=True,
         exists_skip=False,
+        force_download=True,
     ):
         """download cn qlib data from remote
 
@@ -188,6 +236,11 @@ class GetData:
         # get 1min data
         python get_data.py qlib_data --name qlib_data --target_dir ~/.qlib/qlib_data/cn_data_1min --interval 1min --region cn
         When this command is run, the data will be downloaded from this link: https://qlibpublic.blob.core.windows.net/data/default/stock_data/v2/qlib_data_cn_1min_latest.zip?{token}
+
+        # get qlib_bin data
+        python get_data.py qlib_data --name qlib_bin --target_dir ~/.qlib/qlib_data/cn_data
+        When this command is run, the data will be downloaded from this link: https://github.com/chenditc/investment_data/releases
+
         -------
 
         """
@@ -198,14 +251,8 @@ class GetData:
             )
             return
 
-        qlib_version = ".".join(re.findall(r"(\d+)\.+", qlib.__version__))
+        self.version = version
+        self.interval = interval
+        self.region = region
 
-        def _get_file_name_with_version(qlib_version, dataset_version):
-            dataset_version = "v2" if dataset_version is None else dataset_version
-            file_name_with_version = f"{dataset_version}/{name}_{region.lower()}_{interval.lower()}_{qlib_version}.zip"
-            return file_name_with_version
-
-        file_name = _get_file_name_with_version(qlib_version, dataset_version=version)
-        if not self.check_dataset(file_name):
-            file_name = _get_file_name_with_version("latest", dataset_version=version)
-        self.download_data(file_name.lower(), target_dir, delete_old)
+        self.download_data(name.lower(), target_dir, delete_old, force_download)
